Fix example build
