--- examples/esp32/WebSocketClientSSLBundle/include/secrets.hpp.template
@@ -1,2 +1,0 @@
-#define SSID "WIFI SSID"
-#define PASS "WIFI PASSWORD"
