+++ examples/esp32/WebSocketClientSSLBundle/platformio.ini
@@ -16,9 +16,10 @@ monitor_speed = 115200
 upload_speed = 921600
 build_flags = 
     -DCORE_DEBUG_LEVEL=5
+lib_deps = ../../../src
 
 extra_scripts = 
     pre:run_gen_script.py
 
 board_build.embed_txtfiles = 
-    x509_crt_bundle
+    data/cert/x509_crt_bundle.bin
